feat: expose Feedback button
