+++ packages/chat/src/components/FeedbackButton/FeedbackButton.component.tsx
@@ -5,15 +5,9 @@ import { RuntimeStateAPIContext } from '@/contexts';
 import { createColorPalette } from '@/styles/colors';
 
 import { activeBackground, activeIconColor, feedbackButtonStyles, iconStyle } from './FeedbackButton.css';
+import type { IFeedbackButton } from './FeedbackButton.interface';
 import { ThumbsDownIcon } from './ThumbsDownIcon.component';
 import { ThumbsUpIcon } from './ThumbsUpIcon.component';
-
-export interface IFeedbackButton {
-  onClick: () => void;
-  variant?: 'up' | 'down';
-  active?: boolean;
-  testID?: string;
-}
 
 export const FeedbackButton: React.FC<IFeedbackButton> = ({ variant = 'up', active, onClick, testID }) => {
   const {
