+++ packages/chat/src/components/FeedbackButton/FeedbackButton.interface.ts
@@ -1,0 +1,6 @@
+export interface IFeedbackButton {
+  onClick: () => void;
+  variant?: 'up' | 'down';
+  active?: boolean;
+  testID?: string;
+}
