+++ packages/chat/src/components/FeedbackButton/index.ts
@@ -1,0 +1,1 @@
+export { FeedbackButton } from './FeedbackButton.component';
