+++ packages/chat/src/ui.ts
@@ -1,2 +1,3 @@
 export * as Avatar from './components/Avatar';
+export { FeedbackButton } from './components/FeedbackButton';
 export { Howdy } from './components/Howdy';
